bin/requirements.txt
gitpython >= 3.1.27